refactor(SwingSet): unify vat-worker filenames
fixes #2202
