+++ packages/SwingSet/src/kernel/vatManager/manager-nodeworker.js
