+++ packages/SwingSet/src/kernel/vatManager/supervisor-nodeworker-cjs.js
@@ -9,4 +9,4 @@
 
 // eslint-disable-next-line no-global-assign
 require = require('esm')(module);
-module.exports = require('./nodeWorkerSupervisor');
+module.exports = require('./supervisor-nodeworker');
