+++ packages/SwingSet/src/kernel/vatManager/supervisor-nodeworker.js
