+++ packages/SwingSet/src/kernel/vatManager/supervisor-subprocess-node.js
